LayoutTests/accessibility/ios-simulator/heading-with-aria-label-expected.txt

visible text visible text visible text
text1link

This tests that if heading has aria-label, it will override the visible text as necessary.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS heading1Text.traits == heading2Text.traits is true
PASS heading3Text.traits == heading2Text.traits is true
PASS heading1Text.description is 'AXLabel: test label'
PASS heading2Text.description is 'AXLabel: visible text'
PASS heading3Text.description is 'AXLabel: hidden label'
PASS heading4Text1.description is 'AXLabel: text1'
PASS linkText.description is 'AXLabel: link'
PASS successfullyParsed is true

TEST COMPLETE
